css/css-viewport/zoom/discrete-animation.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Zoom: animation and transition tests</title>
<link rel="help" href="https://drafts.csswg.org/css-transform-2/#propdef-zoom">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>

<script>
test_no_interpolation({
  property: 'zoom',
  from: '1',
  to: '2',
});
</script>
